chore: Add compiled JS
